+++ team3project/src/main/java/pack/admin/service/AdminService.java
@@ -5,14 +5,14 @@ import java.util.Optional;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pack.admin.model.AdminUpdateDto;
 import pack.admin.domain.Admin;
-import pack.newbook.domain.NewBook;
-import pack.oldbook.domain.OldBook;
-import pack.common.enums.Grade;
-import pack.newbook.model.NewBookDto;
+import pack.admin.model.AdminUpdateDto;
 import pack.admin.repository.AdminRepository;
+import pack.common.enums.Grade;
+import pack.newbook.domain.NewBook;
+import pack.newbook.model.NewBookDto;
 import pack.newbook.repository.NewBookRepository;
+import pack.oldbook.domain.OldBook;
 import pack.oldbook.repository.OldBookRepository;
 
 @Service
@@ -147,10 +147,7 @@ public class AdminService {
 
     public void removeOb(int obNo) {
         final Optional<OldBook> optionalOldBook = oldBookRepository.findById((long) obNo);
-
-        optionalOldBook.ifPresent(oldBook -> {
-            oldBookRepository.delete(oldBook);
-        });
+        optionalOldBook.ifPresent(oldBookRepository::delete);
     }
 
 }
