+++ team3project/src/main/java/pack/inquery/repository/InqueryRepositorySupportImpl.java
@@ -1,6 +1,6 @@
 package pack.inquery.repository;
 
-import static pack.domain.entity.QInquery.inquery;
+import static pack.inquery.domain.QInquery.inquery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -8,27 +8,28 @@ import pack.inquery.domain.Inquery;
 
 @RequiredArgsConstructor
 public class InqueryRepositorySupportImpl implements InqueryRepositorySupport {
+
     private final JPAQueryFactory jpaQueryFactory;
 
     @Override
     public List<Inquery> findAllOrderByInqOnumASC() {
         return jpaQueryFactory.selectFrom(inquery)
-                .orderBy(inquery.inqOnum.asc())
-                .fetch();
+            .orderBy(inquery.inqOnum.asc())
+            .fetch();
     }
 
     @Override
     public Long getMaxNum() {
         //coalesce 는 nvl의 역할 null일 경우에 해당 값으로 채움
         return jpaQueryFactory.select(inquery.inqNo.max().coalesce(0L))
-                .from(inquery)
-                .fetchOne();
+            .from(inquery)
+            .fetchOne();
     }
 
     @Override
     public Long updateOnum(int inqOnum, int inqGnum) {
         return jpaQueryFactory.update(inquery).set(inquery.inqOnum, inquery.inqOnum.add(1))
-                .where(inquery.inqOnum.goe(inqOnum).and(inquery.inqGnum.eq(inqGnum))).execute();
+            .where(inquery.inqOnum.goe(inqOnum).and(inquery.inqGnum.eq(inqGnum))).execute();
     }
 
 }
