+++ team3project/src/main/java/pack/user/controller/ListController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 import pack.newbook.domain.NewBook;
-import pack.user.service.NewBookService;
+import pack.newbook.service.NewBookService;
 
 @Controller
 @RequiredArgsConstructor
