+++ team3project/src/main/java/pack/user/controller/SearchController.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
-import pack.user.service.NewBookService;
+import pack.newbook.service.NewBookService;
 
 @Controller
 @RequiredArgsConstructor
